Update the access to gradebook to allow course staff and admins
